refactor 1
