+++ .eslintrc.json
@@ -14,7 +14,7 @@
     "ecmaVersion": 2018
   },
   "rules": {
-    "semi": [2,"always"],
+  
 
     "camelcase": 2,
 
@@ -24,7 +24,7 @@
 
     "indent": 2,
 
-    "quotmark": "single",
+
 
     "no-console" : 0,
